ValidatingEditbox rename to FormInput
