+++ src/lib/index.ts
@@ -11,4 +11,4 @@ export { default as ButtonIcon } from './components/Button/ButtonIcon.svelte'
 export { default as ButtonMenu } from './components/Button/ButtonMenu.svelte'
 export { default as Checkbox } from './components/Checkbox/Checkbox.svelte'
 export { default as Editbox } from './components/Editbox/Editbox.svelte'
-export { default as ValidatingEditbox } from './components/Editbox/ValidatingEditbox.svelte'
+export { default as FormInput } from './components/Editbox/FormInput.svelte'
